requirements.txt
boto3==1.14.25
botocore==1.17.25
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
connexion==2.7.0
docutils==0.15.2
Flask==1.1.2
idna==2.10
inflection==0.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==2.6.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
openapi-spec-validator==0.2.8
protobuf==3.12.2
PyJWT==1.7.1
python-dateutil==2.6.0
python-dotenv==0.14.0
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
swagger-ui-bundle==0.0.8
urllib3==1.25.9
Werkzeug==1.0.1
